calendly/meeting/serializers.py
from rest_framework import serializers
from .models import Meeting


class MeetingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Meeting
        fields = ['user1', 'user2', 'start_time', 'end_time']
